+++ docs/src/content/en/reference/observability/logger.mdx
@@ -57,3 +57,121 @@ logger.error("An error occurred"); // Logged as ERROR
 />
 
 **Note:** Some loggers require a `BaseLogMessage` object (with `message`, `destinationPath`, `type` fields). For instance, the `File` and `Upstash` loggers need structured messages.
+
+## File Transport (Structured Logs)
+
+```typescript showLineNumbers copy
+import { FileTransport } from "@mastra/loggers/file";
+
+const fileLogger = new PinoLogger({
+  name: "Mastra",
+  transports: { file: new FileTransport({ path: "test-dir/test.log" }) },
+  level: "warn",
+});
+
+fileLogger.warn("Low disk space", {
+  destinationPath: "system",
+  type: "WORKFLOW",
+});
+```
+
+## Upstash Logger (Remote Log Drain)
+
+```typescript showLineNumbers copy
+import { UpstashTransport } from "@mastra/loggers/upstash";
+
+const logger = new PinoLogger({
+  name: "Mastra",
+  transports: {
+    upstash: new UpstashTransport({
+      listName: "production-logs",
+      upstashUrl: process.env.UPSTASH_URL!,
+      upstashToken: process.env.UPSTASH_TOKEN!,
+    }),
+  },
+  level: "info",
+});
+
+logger.info({
+  message: "User signed in",
+  destinationPath: "auth",
+  type: "AGENT",
+  runId: "run_123",
+});
+```
+
+## Parameters
+
+<PropertiesTable
+  content={[
+    {
+      name: "type",
+      type: "CONSOLE" | "FILE" | "UPSTASH",
+      description: "Specifies the logger implementation to create.",
+    },
+    {
+      name: "level",
+      type: "LogLevel",
+      isOptional: true,
+      default: "INFO",
+      description: "Minimum severity level of logs to record. One of DEBUG, INFO, WARN, or ERROR.",
+    },
+    {
+      name: "dirPath",
+      type: "string",
+      isOptional: true,
+      description: 'For FILE type only. Directory path where log files are stored (default: "logs").',
+    },
+    {
+      name: "url",
+      type: "string",
+      isOptional: true,
+      description: "For UPSTASH type only. Upstash Redis endpoint URL used for storing logs.",
+    },
+    {
+      name: "token",
+      type: "string",
+      isOptional: true,
+      description: "For UPSTASH type only. Upstash Redis access token.",
+    },
+    {
+      name: "key",
+      type: "string",
+      isOptional: true,
+      default: "logs",
+      description: "For UPSTASH type only. Redis list key under which logs are stored.",
+    },
+  ]}
+/>
+
+Custom Transport
+
+You can create custom transports using the `createCustomTransport` utility to integrate with any logging service or stream.
+
+## Example: Sentry Integration
+
+```typescript showLineNumbers copy
+import { createCustomTransport } from "@mastra/core/loggers";
+import pinoSentry from 'pino-sentry-transport';
+
+const sentryStream = await pinoSentry({
+  sentry: {
+    dsn: 'https://c9511f0d57a24887efab1e17aaf15993@o515228.ingest.us.sentry.io/4509564877144069',
+    _experiments: {
+      enableLogs: true,
+      beforeSendLog: (log) => {
+        console.log('beforeSendLog', {log})
+        return log
+      },
+    },
+  },
+});
+
+const customTransport = createCustomTransport(sentryStream);
+
+const logger = new PinoLogger({
+  name: "Mastra",
+  transports: { sentry: customTransport },
+  level: "info",
+});
+```
